Consolidate shared HALs and isolate HAL compile
For realz this time yo
