+++ platformio.ini
@@ -24,7 +24,7 @@ boards_dir  = buildroot/share/PlatformIO/boards
 env_default = megaatmega2560
 
 [common]
-default_src_filter = +<src/*> -<src/config>
+default_src_filter = +<src/*> -<src/config> -<src/HAL> +<src/HAL/shared>
 build_flags = -fmax-errors=5
   -g
   -ggdb
